+++ CHANGELOG.md
@@ -9,3 +9,8 @@ All notable changes to this project will be documented in this file.
 ## Version 1.0.1
 
 - Added officials
+
+## Version 1.1.0
+
+- Added C/C++ source and header file generation commands.
+- Added Contribution section